pyproject.toml
[tool.poetry]
name = "lmsnew"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.30.0"
coverage = "^7.6.9"
drf-yasg = "^1.21.8"
stripe = "^11.4.1"
eventlet = "^0.38.2"
celery = "^5.4.0"
redis = "^5.2.1"
django-celery-beat = "^2.7.0"
gevent = "^24.11.1"
pybloom-live = "^4.0.0"
black = "^24.10.0"
isort = "^5.13.2"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"


mypy = "^1.13.0"
[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

black = "^24.10.0"
[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''


isort = "^5.13.2"
[tool.isort]
# максимальная длина строки
line_length = 119


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
